♻️ :: main.dart 불필요한 코드 이동
menu_bottom.dart 파일 생성, 해당 파일로 네비게이션 코드 이동, 메인에서 간단한 호출로 네비게이션 바 불러오기
