+++ lib/screens/login_screen.dart
@@ -148,7 +148,7 @@ class _LoginScreenState extends State<LoginScreen> {
     if (user.isNotEmpty) {
       Navigator.pushReplacement(
         context,
-        MaterialPageRoute(builder: (context) => const HomeScreen()),
+        MaterialPageRoute(builder: (context) => const MainPage()),
       );
     } else {
       setState(() {
